backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validation and settings
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-dotenv==1.0.0

# HTTP client for scanning
httpx==0.25.2
aiohttp==3.9.1

# DNS and network discovery
dnspython==2.4.2

# SSL/TLS analysis
cryptography==41.0.7
pyOpenSSL==23.3.0

# HTML parsing (for Wappalyzer)
beautifulsoup4==4.12.2
lxml==4.9.3

# Async task processing
celery[redis]==5.3.4
redis==5.0.1

# Rate limiting
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
